Synchronizing Servos Test Code
This is a python program that is used for testing ideas about synchronization between servos. Currently it has code that will synchronize the servos to move together. I need to add the functionality of having the servos stop and wait for the dynamic finger servo to open drop the coin and then the servos can travel back to the initial position