Update about Indentation error inside comment block
For some reason this code has a problem with indentation that occurs inside a comment block. I don't understand why or how. My current solution to run this code without this error is to type all of it again inside a new file before running it. if that works then this current version of the code will be replaced with the working one