Fixed Error of multiple prints from servo()
If there was sufficient time delay between the servo's all finishing and the ServoComplete() from clearing servo_start then the servo function would print 2 values for each letter, which is not correct. Therefore I added a servo_pause.wait() to stop that from happening and a clear and set for the Event. Also put the code that the servo function used inside a if statement to prevent it from getting stuck in an infinite wait, ie, waiting for servo_start to be set and then for servo_pause to be set so it could exit the loop once it realized that self.run was set to False