added True and False to coin_sensor()
replaced 0 and 1 with True and False inside coin_sensor() and changed the naming convention of coin_sensor to match the naming convention of the other code.

Added a comment to a line that needs to be made general. The comment is inside the stop_servos() function and is about the IR_counter line and how the function references the IR_counter.