made sync_servo class code more general
I added an input object for the run_servos and stop_servos functions, to make it more general.

Tested this code on the PI and it works. However the program crashes as we are aware of when all the times in the coin_queue have been removed. The coin_queue will never be empty when the IR sensor gets tripped, by definition that is impossible. 

To solve this a function will have to be added to the sync_servos() class that set self.run to false once the machine wants to be shutdown, by someone pressing a button.