Added code for Finger servo
The added code synchronizes the servo finger with the other servos of the arm to allow the arm to pick up coins and drop coins at the correct time. To do this another function was added inside the sync_servos class called servo_finger. 4 more class variables were added to synchronized the picking up and dropping of the coin. the 4 class variables were 2 counters for the arm servos and 2 Boolean values for the finger servo, one of each for picking up the coin and dropping the coin.